Screen/Moviewriting: Add a hack that doesn't help, doesn't hurt either.
If one specifies codec options of 'AddAudioTrack AudioSource=blahblah',
one can add an AudioSource according to gst-launch line spec blahblah
for video+audio recording with the DC1394 video capture engine.

In theory at least. In practice i could not find any way to make it
work. It always produces an empty file or a file with audio track
only :(. Reason unclear, but as this works with regular Video capture
from GStreamer, it may be a problem of the specific AudioSource=
spec. Trying with the gst-launch line for the full pipeline also does
not work btw.

So this has been verified to not hurt standard movie writing with/without
sound track (ImagingStereoDemo writemovie=1/2) or standard video
recording from GStreamer with/without audio via VideoRecordingDemo.m,
nor does it hurt pure video recording with DC1394 engine.

It didn't help video+audio with DC1394, but maybe the problem is
somewhere else, so commit this anyway as a baby-step into the right
direction. Or so i hope...