XOrgConfCreator: Update for PTB 3.0.17 and future X-Server 1.21+
The modesetting-ddx of X-Servers after 1.20 does support setup of
Variable refresh rate VRR, and the use of modern GAMMA_LUT's if the
Linux kernel supports them. This allows use of modesetting-ddx in
a VRR scenario, but also on Intel gpu's with a color depth of more
than 8 bpc, where Intel-kms needs use of the GAMMA_LUT instead of
the legacy lut, which could only do 256 slots of 8 bpc width.

Also update specs wrt. color depth and VRR and Intel / AMD gpu's.

Also try to retain modesetting-ddx as driver choide if user does
not request otherwise and no forcing circumstances force a different
choice and the modesetting-ddx is already in use.