windowsmakeit64_twisty: Fix build for PsychOculusVRCore driver.
Different include path on the build machine.

Note that i can not test PsychOculusVRCore anymore on Windows-10, both
because the OS refuses runtime installation, and because i lack suitable Rift
DK1/DK2 hardware. Also, the runtime is no longer available for download on
Oculus website, or anywhere else, so this is just to keep existing user legacy
installations working.

Maybe it is time to drop support for this driver on Windows.