VRInputStuffTest.m: Visualize guardian playbounds / outer bounds.
So this is meaningfully tested as well.

Note one limitation: If we change heading via the mouse, then this will misalign
the visualized bounds with the actual bounds. We'd need to setup our rendering
and transformations differently to avoid this. Also tested is that we can request
permanent guardian walls visualization via the optional 1 flag to 'VRAreaBoundary'.