help GStreamer: Recommend GStreamer 1.18.5 from now on.
1.18.2 has sometimes stability issues on macOS, at least with 10.15.7
and Matlab R2021b or Octave 6.3, although Matlab seems to be more
fragile, as usual.

Note: While 1.18.5 on Windows-10 21H1 with R2021b or Octave 6.3
shows no problems in movie playback/writing, and video capture/recording,
with or without sound, on macOS we see that movie playback, movie writing
with/without sound and videocapture works fine, but videorecording is broken
for unknown reasons -- no known workaround. At least this is true on macOS
10.15.7 on a MBP 2017, using its internal FaceTime HD webcam. Go figure...