PsychPortAudio/Linux: Extend half-duplex capture workaround to v19.7.0.
Portaudio v19.7.0 has the same bug as 19.6, because i never got around
to write and submit an upstream fix.

This becomes relevant now, because Matlab R2021b ships its own override
Portaudio library v19.7 for Linux, with the bug present, but our hack
to workaround the bug was limited to 19.6.0 and earlier.

Extend the workaround to 19.7.0 and earlier. This fixes R2021b in
half-duplex capture mode, ie. timestamping of captured audio is
fine again.