Screen/OSX: Enable PTB kernel driver on Intel iGPU by default.
Use kernel driver for Intel iGPU by default.
setenv('PSYCH_DISALLOW_DANGEROUS', '1'); will disable it.

We never received any reports of trouble with this, neither did i ever encounter
it. So lets optimistically assume the problem is not one in practice.

Maybe it helps users to have less broken Intel graphics, or get an assessment
of how broken it is on macOS. I don't expect Apple to ever release new
machines with Intel graphics, so maybe we get lucky with the existing set.