LoadIdentityClut(): Use default lut 0 for Linux + NVidia.
We know lut 0 is the right one for GTX-1650 with nouveau-kms on
Ubuntu 20.04.3-LTS with Linux 5.13-rc, instead of the currently
proposed type 3. We also know that identity pixel passthrough of
that gpu with the Nvidia proprietary driver was broken a bit.

So lets use the default type 0 lut, which definitely fixes nouveau,
and maybe does something good for the proprietary blob as well,
although i'm too short on time to test this now.