Screen/Linux: Skip "swap to even sbc" workaround for Vulkan display.
Skip this workaround of always ending with an even-numbered swap
count sbc before closing the onscreen window if we use an external
display method, e.g, Vulkan display backend.

The workaround was only needed to work around flaws in some old
X-Servers or ddx'en, so given we now require X-Server 1.20 and later
and other year 2020+ stuff, we could drop it completely, but i'm
sentimental.

This gets in the way of some bug diagnosis with multi-output Vulkan
though, so lets disable it at least for Vulkan.