New MultiWindowVulkanTest(): Test multi-display exclusive operation under Vulkan.
This is currently Linux/X11 only, to exercise PTB and the Vulkan drivers
for multi-display exclusive direct display mode. Iow. RandR output leasing
is exercised. This was used to debug a X-Server 1.20 bug, which is now
fixed in master, should be fixed in upcoming Server 21.1.

Add the test to test against future regressions or on other display
systems.