AssertOpenGL: Add some init hack for setup of Wayland specific Screen.mex.
This gets called by scripts which call AssertOpenGL or call
PsychDefaultSetup(), ie. well-behaved ones.

It will detect if we are running under 64-Bit Linux + Octave on a
Wayland setup. If so, it will prepend the Wayland/ subfolder with
the Screen.mex for 64-Bit Octave5+ on Linux + Wayland to the path,
and set other workarounds, to use the Screen.mex for Wayland instead
of the regular production one for X11 native.