PsychPortaudio/Linux: Enable Pa_IsFormatSupported() check on Linux 5.13+ again.
The relevant bug has been fixed in Linux 5.13 and later, so the issue is
resolved for Ubuntu 21.10 (Linux 5.13), and future Ubuntu 20.04.4-LTS
(Linux 5.13) with eta around Jan/Feb 2022, and future Ubuntu 22.04-LTS
and other autumn 2021 and later distributions.

While it would be possible to also add a workaround to a future Portaudio
v19.8.0 release, implementing the following linked proposal, which would be identical
to the fix i had in mind...

https://github.com/alsa-project/alsa-lib/issues/125#issuecomment-797980267

... it probably doesn't make much sense wrt. distro release timing.
A portaudio 19.8 release would end up in Ubuntu 22.04-LTS the earliest,
but that one will ship with Linux 5.14 or later, so the workaround would
not be needed. And even 20.04.4-LTS will be released early 2022 before
a fix could be backported the earliest.

Therefore, we let time solve this problem for us soon. If running on
Linux 5.13+ we enable the check again, otherwise we skip the check,
just as in the past.

-> Tested on Linux 5.13 and Linux 5.11 on Ubuntu 20.04.3-LTS.