Base/OSX: Try to detect if running on Apple M1 ARM Silicon.
We assume that the OS kernel in use faithfully corresponds to the actual
machine architecture, so use the "RELEASE_ARM" substring in the kernel
version string as proxy for a ARM M1 machine or later.

Why? Because i don't know of a better way, given that things like VM's or
Rosetta2 emulation may hide the true machine architecture from us when
using regular OS api's. Also PTB is built for Intel 64-Bit and only Rosetta
runtime translated to ARM, so compile-time checks would not work.

When we are at it, cleanup formatting of PsychTimeGlue.h. This only tested
on Intel, i don't have access to any M1 SoC and our users are as usual too
lazy to lift a finger for their own good. So guesswork it is, with some help
from Wikipedia to find the kernel architecture substring, hopefully.