Screen/OSX: Try to detect and work around Apple M1 SoC's.
The new ARM M1 detection function is used to find out at 'OpenWindow'
time if we are running on a M1 system with Apple proprietary gpu.

If so, we apply workarounds to prevent complete failure and warn the user
about the upcoming visual stimulation timing disaster and hint at our Vulkan
backen as one possible workaround - as in "untested, but probably can't be
worse than what you get without it".

Background: Apples OpenGL driver for M1 SoC gpu's is an OpenGL emulator
on top of Metal. Seems to work well enough for general use cases, but it
inherits all the restrictions of Metal wrt. actual image presentation, specifically
the only way to present to the display on a Apple proprietary gpu, or with Metal
in general, is via CoreAnimation (CA). CA doesn't work with CGL direct display
fullscreen mode and thereby knocks out our CGL based OpenGL path for high
timing precision. CA only works with Cocoa/NSWindow style windowing and
corresponding OpenGL, something that has completely broken timing though.

So if we detect an M1 SoC, we warn the user about all the evil coming his way,
and switch to our backup Cocoa/NS path to at least get a onscreen window with
OpenGL and broken timing, instead of complete error abort.

-> May make the M1 machines usable for toy/demo use, or may not, who knows?