Python: Update docs and defaults for KbQueueFlush.
The default flush_type 1 which was used, is not a good choice for typical
Python use cases, and also can have some undesired side effect on MS-Windows, see issue #745 at
https://github.com/Psychtoolbox-3/Psychtoolbox-3/issues/745

Lets change the default to 0 to simply query the number of pending
input events, and update docs slightly.

Change is untested by myself, but confirmed to do the right thing according
to testing by Sol Simpson (@isolver) on MS-Windows.

In theory flushType 1 should not affect KbQueueEventGet, and in practice
it doesn't on Linux and macOS. But on Windows it could cause loss of some
input events due to timing races between PsychHID's input processing thread
and event arrival from the OS input drivers and clearing the DirectX/DirectInput
hw event queue - a side effect of a workaround we need to work around Windows
DirectInput quirks.

This is an issue that is not easy to resolve on Windows, and could also cause
trouble for PTB on Octave/Matlab, but likely won't in practice, as users do not
mix KbEventAvail/KbEventGet keyboard queue queries and the older
KbQueueCheck api in reasonable use cases, and KbEventAvail() does its
queries with flushType 0. At least so far, no problems were reported over the
last decade or so...