BasicSoundPhaseShiftDemo.m: Fix repetition calculation code, comment more.
Turns out running at 48000 Hz sampling rate ends badly, because then
nsamples works out to be already an integral number of samples, so the
old nsamples - fix(nsamples) calc ends in zero, which ends in infinite wavedur,
which ends badly.

We need to special-case this, for the case of only 1 period of repetition needed.
Also describe the purpose of those lines of code and alternatives better.