PsychVulkanCore: Use VK_KHR_present_wait/id if supported.
The new VK_KHR_present_id + VK_KHR_present_wait extensions, if supported,
allow to wait for present completion for a present. Make use of these to do a
blocking wait with 1 second timeout for present completion.

This requires a recent Vulkan SDK for support and definitions of the new
extensions, so we now build against the latest 1.2.189.2 SDK on Windows,
the windowsmakeit build script got updated accordingly.

If we have support for the timing extension, this will make the wait more
efficient - a blocking wait for present completion, instead of a speculative
timed wait + a tight polling loop, which creates lots of cpu load.

If we don't have the timing extension, it is - in theory - the proper clean and
correct wait to wait for stimulus onset and therefore aid our self-made time-
stamping, reliable in all cases - supposedly a big improvement over our
home-made method.

In practice, this is currently supported by the latest NVidia proprietary drivers
on Windows and Linux, since about August 2021, for the 470.56.5.0 series on
Linux and the 471.69.0.0 series on Windows. An implementation for the Linux
FOSS graphics stack via Mesa is available, but not yet merged upstream.

Unfortunately, testing under Windows-10 21H1 with a GeForce GTX 1650 shows
this to be utterly broken as of the latest Windows 10 driver 496.49.0.0, so unless
or until this gets fixed by NVidia, it is useless at least on Windows. The problem
seems to be that the present wait function does not wait until true present
completion, ie., pageflip completion, but instead until either the driver has
completed hw flip programming, or the image has been handed off to the
desktop compositor DWM, iow. return from wait is usually at least 1 video
refresh cycle too early. Apparently somebody wanted to check off an item
on a feature todo list, without actually doing the hard work of a proper
implementation and testing, so they half-assed it. Yay!

-> Lets hope this goes better on Linux, test pending.