PsychVulkanCore/Windows: Make sure fullscreen exclusive is forbidden if usercode doesn't want it.
So updating the AMD graphics+display driver to the latest pile of horseshit,
version 21.11.2 30.0.13033.5003, brings us new Vulkan bugs on top of all the
old Vulkan bugs, because nobody nowadays can be arsed to do QA
anymore. This on Windows 10 21H1 with all updates applied.

Now fullscreen exclusive mode is broken to the point where it only displays a
black screen, and vkQueuePresent() failure with error code -1000255000 aka
VK_ERROR_FULL_SCREEN_EXCLUSIVE_MODE_LOST_EXT despite the switching
function reporting successful switch to fs exclusive mode before. Also no keyboard
input is accepted anymore and one has to kill Matlab/Octave to escape the black
screen. This is one step down from having to ALT+Tab in a previous driver
version to escape the black screen, because why stop at merely bad if one can go
to much worse?

We now make sure that if the "disable fs exclusive mode" flags 2 are specified, we
not just merely do not request fullscreen exclusive, but instead explicitely forbid the
driver to switch to fs exclusive.

This gives us a picture back, and working keyboard input, and non-crashing
Matlab/Octave, and as a side-effect broken visual stimulation timing, ofc.

Also, now we run into a new problem in that in non-fs-exclusive mode, HDR needs
to be manually enabled/disabled by the user in the Windows display settings GUI.
and apparently the Windows DWM compositor has forgotten how to properly do
HDR compositing in the first place, and all HDR stimuli look wrong, as if a HDR
signal is sent to the monitor (monitor confirms), but with SDR EOTF instead of a
HDR EOTF.

There's just no winning on Windows... But hey, it worked fine for a year, before
falling apart. What more can one want?