PsychVulkanCore: Fix crash at verbosity 4+ on Linux.
It is not enough to check for fpGetRefreshCycleDurationGOOGLE != NULL,
the current vulkan device also must support the timing extension.

Why? Because we get that function pointer via instance proc address
query, so it is non-NULL if at least one ICD supports timing. Calling
on a non-supporting ICD may end with a NULL-Ptr reference and boom!

-> Happened with our experimental Mesa radv buid with timing support.