PsychVulkan: More hacks to deal with fragile NVidia Vulkan direct display.
After a fricking whole day i came to the conclusion that the NVIDIA
proprietary Vulkan is incredible fragile and unpredictable about when
it manages to lease a RandR output for direct display mode, and when
not. The pattern seems to follow no rhyme or reason.

Therefore lets avoid it whenever possible. We add a check that will
skip direct display mode if an onscreen window on Linux+X11+NVIDIA
covers the whole X-Screen. In that case, just creating a fullscreen
X11 window, opaque, top-level, decorationless etc., will display the
Vulkan content via pageflipping, so timing and timestamping is fine,
no need for the drama and flicker-fest of the NVIDIA driver.

In general, direct display mode does essentially not have any advantages
for us at all. NVIDIA does not support higher than 10 bpc precision SDR
framebuffers, or HDR display, or special G-Sync goodies in direct
display mode. Neither does it allow direct display mode on other
gpu's in an Optimus setup than the NVidia itself. So Vulkan/WSI/X11 is
fine for almost all cases. The only purpose of direct display mode is
convenience, if one has a multi-display setup and does not want to bother
setting up a separate X-Screen. Then the RandR output leasing is a little
bit of convenience - in the cases when it unpredictably works, that is...

The same goes for the Vulkan display backend as a whole. The only use
case atm. apart from above multi-display convenience, would be if it
would allow for proper timing and timestamping in Optimus PRIME render
offload scenarios under the WSI/X11 backend. If that is possible remains
to be seen, but tests so far do not look encouraging :/.