linuxmakeit64/octave: Add libxcb-present as Screen mex dependency.
So future code can directly and explicitely interact with and
utilize the X11/Present extension, instead of just indirectly.

We now depend on the xcb-present package. Generally this is
also a (indirect) dependency of Mesa GLX, or the X-Server or
Vulkan drivers, so we might get away with not explicitely
specifying it in the Debian package dependency descriptors,
although adding it there would surely not hurt.