Screen/Linux/X11: Add initial NVidia Optimus proprietary OpenGL support.
This is meant to be used on modern NVidia gpu's of the Maxwell gpu
family (GeForce 800 series) or later, with which can not really use
the Mesa noveau open-source OpenGL graphics and display driver, due
to the poor performance caused by lack of gpu reclocking support and
other NVidia induced shenigans (signed firmware missing for reclock,
power management, sometimes even gfx hw acceleration).

It triggers in "NVidia Optimus" PRIME muxless render offload mode
with an Intel integrated processor graphics iGPU (or also with AMD
Ryzen APUs, although it is not yet useful for AMD iGPU + NVidia dGPU
configurations with current Linux kernels due to amdgpu-kms driver
limitations). It is only used when the NVidia proprietary OpenGL
driver is in use, as we have the Mesa open-source nouveau OpenGL
driver case perfectly covered since many years.

In such a configuration, the NVidia gpu + proprietary driver will
do the OpenGL stimulus drawing/rendering and post-processing.
The X-Server display gpu and driver, ie., Intel or AMD with Mesa
open-source drivers and Linux open-source kms drivers are used for
presenting and displaying the stimulus to the actual video output/
display, ie. all flip scheduling, timing and timestamping is done
by the open-source drivers with excellent timestamping precision and
reliability/trustworthiness for fullscreen onscreen windows (full
X-Screen covering, unredirected, compositor locked out/bypassed,
kms-pageflipping used).

However, because NVidia's proprietary GLX module is in use in such
a configuration, instead of the standard X-Server GLX module, all
GLX/OpenGL timing and timestamping extensions are disabled, as
NVidia GLX does not support them.

But as I just learned by xtrace'ing and reverse engineering X-Server
+ NVidia OpenGL protocol interactions, the good news is that the
NVidia OpenGL driver uses X11/DRI3/Present for image presentation in
Optimus mode. The driver subscribes to PresentPixmapCompleteNotify
and PresentPixmapIdleNotify events, receives them from the server
in response to each completed presentation and processes them
internally to a limited degree, apparently for sync-handshaking and
swap throttling, using the present completion msc of a previous
present to calculate the PresentPixmap targetmsc for presenting the
next image as targetmsc = previous_msc + 1, or similar. This is
insofar strange as this is not actually needed for the primitive
glXSwapBuffers implementation of NVidia, a hard-coded targetmsc of
zero would do the trick. And regrettable, because it should be
very easy for NVidia to implement the intel swap events extension
trivially on top of this mechanism and all the work they already did,
ie. they are > 90% there for their GLX module. Instead they seem to
discard the information instead of exposing it.

Anyway, we receive this stream of present notify events from the
server, with all the info we need for excellent timestamping. This
experimental initial commit uses the X11/Present extension directly
to subscribe us with a dedicated special event queue per onscreen
window of our own, so we receive copies of all completion events
in our own queue, from which we can dequeue and process them for
our own purpose! We use them for our own homegrown reimplementation
of intel swap events, and do corresponding accounting so we get
feedback for each completed OpenGL bufferswap, if the swap was
successfull, how it was completed (pageflip hopefully!), and the
vblank count msc and timestamp of present completion.

This should give us flip timestamping of identical quality,
trustworthiness, robustness and precision as with the regular
open-source OpenGL drivers for AMD, Intel, RaspberryPi etc.

So far the theory. In practice:

- Test with Videoswitcher+Bits#/RtBox measurement hardware was
  successfull - Timestamping works reliably :)

- This is only a first beta implementation, so here by dragons!

- Only intel-kms and thereby Intel integrated gpu's can currently
  pageflip and scanout from GTT, so pageflipping and good timestamps
  on Intel only by default.

- amdgpu-kms currently only allows GTT scanout/pageflip when the
  display gpu is a AMD stoney, carrizo or Ryzen APU and the dGPU
  for renderoffload is a modern amdgpu-kms driven AMD gpu. It does
  not allow pageflip with other dGPU's from NVidia or future Intel
  DG-x, or on older AMD APU's (also not on RavenRidge 1st gen), or
  from AMD dGPU's as display gpu's. I do have some kernel patches
  to fix this, which work nicely here on darlene with 1st gen
  RavenRidge (Ryzen 5 2400G, Vega11 gfx, DCN-1 display engine) +
  NVidia Turing family GeForce GTX 1650 under both noveau and
  NVidia proprietary. However these patches implement dirty hacks,
  so they may not be upstreamable, we'll see.

  For AMD or similar, the only workable approach wrt. quality is
  going through the X11 desktop compositor, so no pageflip/direct
  GTT scanout. But this new timestamping method does not mesh well
  with desktop composition, so no-go for timing. However, this new
  approach got me thinking. Maybe we can play some more complex games
  with pixmap present complete, pixmap idle events and somehow get
  present notify events for the compositors main framebuffer window,
  to ordering and timing info for when the compositors is done latching
  our OpenGL frontbuffer, done compositing, and done pageflip presenting
  the composited desktop? Tricky, but not impossible. Maybe a worth
  a try in the future, but not now...