Screen/Linux/X11: Retire use of _KDE_NET_WM_BLOCK_COMPOSITING - redundant.
The _KDE_NET_WM_BLOCK_COMPOSITING atom hint was used in the past to
tell KDE's KWin compositor to get out of the way and disable desktop
compositing while an opaque, decoratinless, top-level, unoccluded
PTB fullscreen onscreen window was open, to allow use of pageflipping
directly. This was done as an automated way to deal with this, while
there weren't alternatives. KDE supported this since July 2011.

Then _NET_WM_BYPASS_COMPOSITOR came along as an official cross-desktop
NetWM standard with slightly higher flexibility, first implemented in
GNOME afaik since December 2012.

With KDE 5.20, support for _NET_WM_BYPASS_COMPOSITOR was introduced on
29. December 2015, in a way that is compatible with the old KDE approach
and overrides that old setting if both are applied - ie. in our case -
so the setting was redundant for all distributions with KDE versions
released in the year 2016 or later.

As we only support Ubuntu 20.04-LTS and later now, we can be sure that
any KDE version supported by us will have a fully sufficient implementation
of _NET_WM_BYPASS_COMPOSITOR, and the old hint is useless.

Therefore remove it.