LoadIdentityClut: Warn about use of modesetting-ddx on modern AMD.
The modesetting-ddx of X-Server 1.20 and earlier can not use atomic
GAMMA_LUT, and so can't set a LUT that triggers identity pixel
passthrough. Only amdgpu-ddx/radeon-ddx/ati-ddx can.

Therefore give a warning and advise to switch to amdgpu et al. if
use of modesetting-ddx is detected. This would have saved me from
a few hours of confusion...