Screen/Linux/X11: Add some diagnostics to swap_events setup.
Print a warning if they can't be enabled. Not that it helped in
our situation, but the code is written, so commit it anyway.

So apparently, the X-Servers GLX implementation does not longer
emit INTEL_swap_events, as soon as NVidia's proprietary GLX module
is loaded on the client side, even though events are reported as
enabled, and even on X-Screens with a non-Nvidia gpu :( Why? Beats
me. The NVidia proprietary driver is a gift that keeps giving...

Anyway, one workaround exists: Explicitely asking for use of the
open-source 'mesa' GLX module via:

setenv('__GLX_VENDOR_LIBRARY_NAME','mesa');

Ofc. this will now block use of OpenGL on the NVidia proprietary
driven X-Screen, so one has too choose ones poison on multi-vendor
multi-gpu setups -- Or simply stick to nouveau, or non-NVidia, as
recommended.