PsychHDR: Make it safe to call from within Screen('BeginOpenGL').
Need to use the type 7 Screen('GetWindowInfo') query, which is safe
to call / allowed from within userspace OpenGL mode.

Note we can not do this for the special Linux/X11 stereo static HDR
hack path, as we do need the devices GLUUID, and that can not be
queried in safe mode 7. Users will have to use 'EndOpenGL' for such
a call into PsychHDR if they really need this.