PsychPortAudio/Linux: Improve default device selection.
For latencymode 0 = High latency, no timing precision, first try to
use PulseAudio's default devices, then JACK's, then ALSA, instead of
only ALSA. This allows collaboration with other audio clients if
PulseAudio is supported in the future, or if JACK is installed and
running now.

For latencymode 1, low-latency, some timing precision, try to play
nice with other clients: Choose PulseAudio, then JACK, then fallback
to ALSA.

For latencymode 2+, exclusive low-latency, go to ALSA as in the past,
for exclusive sound card access with best latency and timing.

This mimicks behaviour as on MS-Windows and macOS.