InitializePsychSound(): Add support for 'reallyneedlowlatency' flag on Linux.
Add some checks to no-op redundant calls to InitializePsychSound() in a
session. Add some consistency check for matching 'reallyneedlowlatency'
flag on successive no-op invocations of InitializePsychSound().

Add the switching logic for Linux which will decide about the
PsychPortAudio('EngineTunables', ...autosuspend) setting for running
PulseAudio servers, depending on the 'reallyneedlowlatency' flag.

If 'reallyneedlowlatency' is set to zero or left out, so it defaults to
zero, then the user script will not need lowest latency, ergo we can
use the PulseAudio backend if it is supported, for sharing of audio
devices among clients. If 'reallyneedlowlatency' is one, then it is
a low-latency sound server like JACK or ALSA, nothing else, and we must
auto-suspend PulseAudio servers, just as we did always on Linux so far.

The switching logic is tested on Linux against current PortAudio 19.6
and 19.7 without PulseAudio support, and also against the development
prototype of a Portaudio with PulseAudio support. It does the right thing.

However, i will switch this switching logic off for the time being in
a followup commit, as we only want to enable this magic once we've tested
the final public release of PulseAudio enabled Portaudio and can be sure
it meets minimum low latency and timing requirements for standard mode.
The current prototype seems to do a good job, but its implementation may
still change for the better or worse. And without access to my Datapixx
atm. i can't run the most stringent timing and latency tests atm., only
less bullet-proof tests.