New bash script coldlinuxbuildOnMatlab.sh: Build mex files without Matlab license.
This script contains the compiler/linker/mv calls that our Matlab build
script linuxmakeit64.m would execute when called as linuxmakeit64(-1) for
rebuilding all our 64-Bit Matlab mex files on 64-Bit Linux on Intel, as
done with Matlab R2021b on Ubuntu 20.04.3-LTS on one of my build machines.

It is based on the "diary on; linuxmakeit64(-1); datapixxmakemex; diary off"
command sequence recorded in the diary log file fulllinuxbuildlogR2021b.txt,
when adding the options '-silent -n' to the mex command, so mex does a
simulated run where it doesn't actually do anything but just prints what
it would do, ie. which compiler/linker calls it makes etc.

Why do we have this? So we are able to still rebuild PTB mex files for
64-Bit Matlab on Linux even if we temporarily don't have access to a
proper Matlab license (-server), so we can't actually launch Matlab and
run the linuxmakeit64 build script.

This allows to keep minimal maintenance up in that situation. We can
rebuild mex files, but not test them on Matlab. However we can build and
test mex files on Octave at all times, and behaviour should be identical
in almost all cases that matter.