Screen: Add some visual indication of some types of triple-buffering.
This will cause some "jiggle" movement of the startup splash image if
the operating system does bypass the desktop compositor, but the display
driver still employs triple buffering. Not all types of triple-buffering
or n-buffering with n > 2 can be detected, e.g., it is expected to not be
diagnostic at all in desktop compositor situations, "Optimus" style render
offload etc., but the most straightforwad type of triple buffering may be
detectable, e.g., what the Intel driver on MS-Windows 10+ uses.

There should be no visible change for double-buffered systems, only on
triple buffering in some conditions. The test has a high false negative
rate. We don't use this on Linux, as we have much better diagnostic built
in since many years due to superior api's for this stuff on Linux. Also,
n-buffering is nothing bad on Linux with FOSS drivers, but often beneficial
for performance.

-> Test if this actually detects Intel triple buffering on Windows is
   pending. So far only confirmed to not be a troublemaker on Linux.

This test was inspired by TripleBufferCheck.m from user _qx1147 on the
PTB forum, thanks. See thread here:
https://psychtoolbox.discourse.group/t/error-synchronization-failure/4176/5