Screen/Linux/X11: Improve NetWM composition handling for NVidia Optimus.
For the typical hybridGraphics == 5 case of using a NVidia dGPU in
combination with an AMD iGPU for Prime renderoffload aka NVidia Optimus
while using the NVidia proprietary graphics driver and GLX module, we
can do a bit better.

Such a combo with AMD iGPU can not scanout from GTT directly, ergo no
direct pageflip possible, so we have to fallback to use of the compositor
as middle-man and translator between the iGPU and dGPU, and ideally use
NetWM for it. Indeed we are limited with the proprietary driver, but
we can do a bit better than nothing:

We do get high precision true flip completion timestamps from the NetWM
compositor which mark the end of a vblank with high precision. Use these
from a previous flip as baseline for translating the tWhen target time
into a delta wait time to wait so we wake up and submit a composition
request exactly at the end of the vblank preceeding the wanted vblank,
about ~2 msecs before the composition deadline, so our request gets
latched for the right composition + flip cycle of the X11 compositor and
we hopefully hit the right frame - not too early, but usually also not
too late. How well it works depends as always on compositor load, ie.
what other clients do rendering in that frame and add gpu work, and how
complex is the scene-graph to composite. But at least we do as good as
we can, given the proprietary blobs imposed limitations.

-> Not tested on actual AMD + NVidia setup yet, but simulation on my
   Intel tablet looks promising.