Screen: Fix MSAA without imaging pipeline by better AUX buffer fallback.
Turns out most gpu + driver combos nowadays do no longer support AUX
buffers, and so trying a context with AUX buffers usually fails. Our
old fallback handling on Linux and Windows would first try to disable
MSAA and only later AUX buffers, which is the wrong order. This will
needlessly disable MSAA when disabling AUX buffers is the solution.

Switch the order of fallbacks on Linux/X11/GLX and Windows, so AUX
buffers are the first requirement to go out of the window.

-> Fixes "no MSAA without imaging pipeline" problem on Linux with Mesa
   OpenGL for at least AMD and Intel, probably all other drivers as well.

The same logic may help on MS-Windows. macOS OpenGL implementation still
seems to support AUX buffers and as it is frozen in time, will probably
until the bitter end, so no need for fallback order changes there.

When we are at it, fix a warning for use of GL_MULTISAMPLE_FILTER_HINT_NV
on implementations that don't support GL_NV_multisample_filter_hint, and
make multiSample enable consistent to work for any value > 0, instead of
> 1.

Why the ruckus now? We want line smoothing with proper endcaps on AMD
with Mesa 22+, and that requires MSAA, and it would be nice if it would
not also require full imaging pipeline unneccessarily, as there is
overhead associated with both.