PsychPortAudio/Linux: Change realtime priority setup for ALSA and PulseAudio.
The old setup only used RT_FIFO realtime scheduling for the audio
callback thread with priority 1, which is the lowest possible RT
level and the same as the PTB mainthread if Priority(1) is used.

This may be too little on heavily loaded machines with other RT
threads active. See the following reported issue on a MS-Surface Pro 6
under Ubuntu 20.04 LTS:

https://psychtoolbox.discourse.group/t/audio-and-surface-pen-questions/4276

The issue was not reproducible locally, but pointed to this potential
weakness in our RT scheduling setup.

New approach: Ignore the Portaudio method for RT setup and use our own
PsychSetThreadPriority() function for a one-time setup on first paCallback
invocation to switch to RT_FIFO scheduling and boost the RT priority to
4 levels above the initial startup level. As initial level and scheduling
is inherited from the PTB mainthread, this effectively turns into the
following behaviour:

1. Try RT_FIFO with RT priority = RT Priority() of main-thread + 4.
2. On failure (due to lacking permission to raise prio, print a warning
   and stick to current inherited priority, which could still be RT if
   the main-thread used a Priority(n) with n > 0.

Note: Our default setup (from PsychLinuxConfiguration() durint setup) will
try to allow for RT priorities up to level 50, so this boosting should
work for main-thread Priority up to Priority(46) (46+4 <= 50).

For a typical Priority(MaxPriority()) level of Priority(1), this will
boost to RT priority 1+4 = 5, which is higher than the main thread and
higher than all other current PTB background processing threads, which
usually boost +1 or +2.

Lets see if this resolves the issue reported above.