linuxmakeit64: Add support for CI build on GitHub.
Small mods to make the build on the GitHub-Actions
Linux runner for Matlab work, so we can build the
64-Bit Matlab Linux mex files for Intel.

The new optional 'cibuild' parameter marks a build on
the CI system if set to 1. It omits building the
Eyelink mex file, which we can't build atm. without
the eyelink sdk inside the CI instance.