Setup/Download/UpdatePsychtoolbox(): Allow optional non-interactive operation.
If the new optional parameter 'tryNonInteractiveSetup' is provided with
a non-zero (true) value, then try to perform the whole installation and
setup procedure non-interactive, ie., without user interaction. The
default setting is interactive mode, just like before - the safe choice.

For this purpose, output paging is disabled via "more off", pause()
commands are turned into no-ops via "pause off", and input() commands
asking the user questions are skipped - instead an answer is assumed
which leads to forward progress of the setup procedure, even if this
may lead to destructive operations that may lose data, e.g., by
deleting old Psychtoolbox folders or changing Matlab/Octave settings.

This mode is experimental and essentially untested, so may contain
bugs. It is also mostly unsupported in case of any trouble.

There is real risk of data loss and screwed up system installations,
or of the user simply missing important feedback about errors/warnings
or problems on the host system wrt. PTB installation.

This as a convenience for people that know what they are doing and
what risks may be involved.

This setup method is expected to work (but not actually tested!) on
MS-Windows and macOS. It would likely fail to run non-interactive on
Linux, because the Linux setup function PsychLinuxConfiguration() is
not yet prepared for non-interactive use at all. That setup routine
does request admin passwords under some circumstances, so can't run
completely non-interactive, and running non-interactive may not be
desirable, given that some of the offered setup actions and system
tweaks can be quite intrusive, so maybe users should give some form
of informed consent.

-> This feature may be removed in the future on any sign of trouble.