Screen/Linux/X11: Protect against execution under Wayland.
Trying to run our standard X11/GLX Screen() under a Wayland server, ie.
via XWayland, will end badly - with a hard hang. This due to deficiencies
and bugs in XWayland - a broken OpenML implementation, as of May 2022.

Detect this, and error abort, providing the user with troubleshooting
instructions, ie. the advice to login to a classic X11 native session.

-> Tested on Ubuntu 22.04-LTS with XWayland 22.1.1.
-> On Ubuntu 22.04-LTS + GNOME 42 under Octave, our Screen.mex file
   for Octave + Wayland works ok'ish, at least for non-Retina displays.
   Not production ready, but at least testable. The libwaffle shipping
   with Ubuntu 22.04-LTS is good enough for our current purposes.